% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{cents2ratio}
\alias{cents2ratio}
\title{Convert cents to a ratio}
\usage{
cents2ratio(cents)
}
\arguments{
\item{cents}{Value in cents}
}
\value{
The corresponding pitch ratio
}
\description{
Convert cents to a ratio
}
