% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{chord_midi_file}
\alias{chord_midi_file}
\title{Create Chord MIDI File}
\usage{
chord_midi_file(
  chord,
  keyboard_map,
  output_directory = "~/MIDI",
  file_name = paste0("chord_", paste(chord, collapse = "_")),
  tempo = 60,
  hold_beats = 1
)
}
\arguments{
\item{chord}{a numeric vector with the scale degrees for the chord}

\item{keyboard_map}{the keyboard map for the scale}

\item{output_directory}{character, default "~/MIDI". This will be created
if it does not exist.}

\item{file_name}{character, default is "chord_\code{paste(chord, sep="_")}.mid"}

\item{tempo}{numeric, beats per minute, default is 60}

\item{hold_beats}{numeric number of beats to hold the chord, default is 1}
}
\value{
the full path to the file
}
\description{
Creates a MIDI file to play a given chord on a remapped
synthesizer
}
\details{
The created MIDI file will play all inversions of the chord from
the lowest to the highest notes in the keyboard map.
}
\examples{
\dontrun{
  eikosany <- cps_scale_table()
  eikosany_map <- keyboard_map(eikosany)
  print(chord_midi_file <- chord_midi_file(
    c(1, 6, 11, 15),
     eikosany_map
  ))
}
}
