% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{chord_sawtooth_WAVs}
\alias{chord_sawtooth_WAVs}
\title{Create Chord Sawtooth WAV Files}
\usage{
chord_sawtooth_WAVs(
  chord,
  keyboard_map,
  lowest_note = 48,
  highest_note = 84,
  output_directory = "~/Multisample",
  duration_sec = 1,
  velocity = 100,
  sample_rate_hz = 48000,
  bit_width = 24
)
}
\arguments{
\item{chord}{a numeric vector with the scale degrees for the chord}

\item{keyboard_map}{the keyboard map for the scale}

\item{lowest_note}{the lowest MIDI note number to use. Default is 48.}

\item{highest_note}{the highest MIDI note number to use. Default is 84.}

\item{output_directory}{character, default "~/Multisample". This will
be created if it does not exist.}

\item{duration_sec}{how long to hold each note, default = 1}

\item{velocity}{MIDI velocity, default = 100, max is 127}

\item{sample_rate_hz}{sample rate in hz, default = 48000}

\item{bit_width}{bit width of samples, default = 24}
}
\value{
the full path to output_directory
}
\description{
Creates a \code{gm} music object for a given chord and keyboard map
}
