% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{chord_synth}
\alias{chord_synth}
\title{Synthesize a Chord}
\usage{
chord_synth(chord, duration_sec, velocity = 100)
}
\arguments{
\item{chord}{a vector of frequencies for the chord}

\item{duration_sec}{total duration of the chord in seconds}

\item{velocity}{MIDI velocity, default = 100, max is 127}
}
\value{
a \code{Wave} object containing the synthesized chord
}
\description{
Creates a \code{Wave} object for a given chord
}
\examples{
\dontrun{
  justmajor7th <- c(1, 5/4, 3/2, 7/4)
  wave <- chord_synth(256*justmajor7th, duration_sec = 10)
  tuneR::play(wave)
}

}
