% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{chord_table}
\alias{chord_table}
\title{Create Chord Table}
\usage{
chord_table(harmonics = c(1, 3, 5, 7, 9, 11), choose = 4, scale_table)
}
\arguments{
\item{harmonics}{a vector of the harmonics to use - defaults to the first
six odd numbers, the harmonics that define the 1-3-5-7-9-11 eikosany.}

\item{choose}{the number of harmonics to choose for each chord -
defaults to 4, to compute tetrads for the eikosany.}

\item{scale_table}{the scale table to use for note number and name lookup}
}
\value{
a list of four items:
\itemize{
\item \code{harmonic_note_numbers}: the harmonic chord expressed as
colon-separated MIDI note numbers
\item \code{harmonic_note_names}: the harmonic chord expressed as
colon-separated note names
\item \code{subharmonic_note_numbers}: the subharmonic chord expressed as
colon-separated MIDI note numbers
\item \code{subharmonic_note_names}: the subharmonic chord expressed as
colon-separated note names
}
}
\description{
Creates a chord table
}
\examples{
\dontrun{

# the defaults yield the 1-3-5-7-9-11 eikosany
eikosany_scale <- scale_table()[["scale_table"]]
print(eikosany_scale)
eikosany_chords <- chord_table(scale_table = eikosany_scale)
print(eikosany_chords)
}
}
