% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{cps_label}
\alias{cps_label}
\title{Generate labels for a combination product set (CPS)}
\usage{
cps_label(harmonics = c(1, 3, 5, 7, 9, 11), choose = 3)
}
\arguments{
\item{harmonics}{A vector of harmonic factors to use - defaults to the first six odd numbers}

\item{choose}{The number of harmonics to choose for each combination - defaults to 3}
}
\value{
A character vector of harmonically-labeled notes (e.g., "1x3x5")
}
\description{
Generate labels for a combination product set (CPS)
}
