% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{create_12edo_scale_table}
\alias{create_12edo_scale_table}
\title{Create 12EDO Scale Table}
\usage{
create_12edo_scale_table()
}
\value{
a \code{data.table} with ten columns:
\itemize{
\item \code{note_name}: the note name (character)
\item \code{ratio}: the ratio that defines the note, as a number between 1 and
2
\item \code{ratio_frac}: the ratio as a vulgar fraction (character). The ratios
for 12EDO are irrational, so this is an approximation.
\item \code{ratio_cents}: the ratio in cents (hundredths of a semitone)
\item \code{degree}: scale degree from zero to (number of notes) - 1
\item \code{base_12EDO}: note name for nearest 12EDO note
\item \code{offset_cents}: offset in cents from \code{base_12EDO}
}
}
\description{
Creates a scale table for 12EDO
}
\examples{
\dontrun{

print(vanilla <- create_12edo_scale_table())

}
}
