% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{create_base_keyboard_map}
\alias{create_base_keyboard_map}
\title{Create Base Keyboard Map}
\usage{
create_base_keyboard_map(middle_c_octave = 4)
}
\arguments{
\item{middle_c_octave}{octave number for middle C. The default is 4, but
other software can use 3 or some other number}
}
\value{
the base keyboard map - what a synth tuned to 12 EDO should have.
This is a data.table with three columns:
\itemize{
\item \code{note_number}: the MIDI note number from 0 through 127
\item \code{note_name}: the note name (character)
\item \code{octave}: octave number. This has an offset defined by parameter
\code{octave_offset}.
\item \code{frequency}: frequency in Hz. By convention, A440 is MIDI note number
69, so this can be computed as 440.0 * 2 ^ ((note_number - 69) / 12)
\item \code{cents}: cents above the default for MIDI note 0. By convention, this
is zero for note number 0.
}
}
\description{
Creates a base keyboard map
}
\examples{
\dontrun{

keyboard_map_c4 <- create_base_keyboard_map()
print(keyboard_map_c4)
keyboard_map_c3 <- create_base_keyboard_map(middle_c_octave = 3)
print(keyboard_map_c3)
}
}
