% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{create_chord_plot}
\alias{create_chord_plot}
\title{Create Chord Plot}
\usage{
create_chord_plot(chord, title_string)
}
\arguments{
\item{chord}{a numeric vector with the scale degrees for the chord}

\item{title_string}{A string to use for a plot title. This will usually
be the chord name from a chord table}
}
\value{
a \code{ggplot2} object that can be printed
}
\description{
Creates a \code{ggplot2} object to plot a chord
}
\examples{
\dontrun{
  print(create_chord_plot(c(1, 6, 11, 15), "1:3:5:7"))
}
}
