% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{create_chord_table}
\alias{create_chord_table}
\title{Create Chord Table}
\usage{
create_chord_table(scale_table, choose)
}
\arguments{
\item{scale_table}{the scale table to use for note number and name lookup}

\item{choose}{the number of harmonics to choose for each chord}
}
\value{
a data.table with four columns:
\itemize{
\item \code{chord}: the chord expressed as colon-separated harmonics. A
subharmonic chord is prefixed with a "~".
\item \code{degrees}: the chord expressed as colon-separated scale degrees
\item \code{chord_index}: the row number of the chord in the combination output
\item \code{is_subharm}: zero if it's harmonic, one if it's subharmonic.
}
The resulting data.table is sorted into harmonic-subharmonic pairs using
\code{data.table::setkey.}
}
\description{
Creates a chord table
}
\examples{
\dontrun{


# compute the tetrads of the 1-3-5-7-9-11 Eikosany
eikosany <- create_scale_table()
print(eikosany_chords <- create_chord_table(eikosany, 4))
}
}
