% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{create_edo_scale_table}
\alias{create_edo_scale_table}
\title{Create EDO Scale Table}
\usage{
create_edo_scale_table(
  note_names = c("C ", "C#", "D ", "D#", "E ", "F ", "F#", "G ", "G#", "A ", "A#", "B ")
)
}
\arguments{
\item{note_names}{a character vector with the names of the notes in the
scale. That's all you need to give it - it can figure out everything
else!}
}
\value{
a \code{data.table} with eight columns:
\itemize{
\item \code{note_name}: the note name (character)
\item \code{ratio}: the ratio that defines the note, as a number between 1 and
2
\item \code{ratio_frac}: the ratio as a vulgar fraction (character). The ratios
for most EDOs are irrational, so this is an approximation.
\item \code{ratio_cents}: the ratio in cents (hundredths of a semitone)
\item \code{interval_cents}: interval between this note and the previous note
\item \code{degree}: scale degree from zero to (number of notes) - 1
\item \code{key_12EDO}: note name for nearest 12EDO note
\item \code{offset_cents}: offset in cents from \code{key_12EDO}
}
}
\description{
Creates a scale table for equal divisions of the octave (EDO)
}
\examples{
\dontrun{

print(vanilla <- create_edo_scale_table()) # default is 12EDO, of course

nn19 <- c(
  "C ",
  "C+",
  "D-",
  "D ",
  "D+",
  "E-",
  "E ",
  "E+",
  "F ",
  "F+",
  "G-",
  "G ",
  "G+",
  "A-",
  "A ",
  "A+",
  "B-",
  "B ",
  "B+")
print(length(nn19))
print(nn19)
print(edo19 <- create_edo_scale_table(nn19))

nn31 <- c(
  "C  ",
  "C+ ",
  "C++",
  "D--",
  "D- ",
  "D  ",
  "D+ ",
  "D++",
  "E--",
  "E- ",
  "E  ",
  "F--",
  "E++",
  "F  ",
  "F+ ",
  "F++",
  "G--",
  "G- ",
  "G  ",
  "G+ ",
  "G++",
  "A--",
  "A- ",
  "A  ",
  "A+ ",
  "A++",
  "B--",
  "B- ",
  "B  ",
  "C--",
  "B++")
print(length(nn31))
print(nn31)
print(edo31 <- create_edo_scale_table(nn31))

}
}
