% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{create_interval_table}
\alias{create_interval_table}
\title{Create Interval Table}
\usage{
create_interval_table(scale_table)
}
\arguments{
\item{scale_table}{a scale table from \code{create_scale_table}}
}
\value{
an interval table. This is a data.table with seven columns
\itemize{
\item \code{from} name of "from" note
\item \code{from_degree} scale degree of "from" note
\item \code{to} name of "to" note
\item \code{to_degree} scale degree of "to" note
\item \code{ratio} interval as a number
\item \code{ratio_frac} interval as a vulgar fraction (character)
\item \code{ratio_cents} interval in cents
}
}
\description{
Creates an interval table from a scale table
}
\examples{
\dontrun{

# default is the 1-3-5-7-9-11 Eikosany
eikosany <- create_scale_table()
print(eikosany_interval_table <-create_interval_table(eikosany))

}
}
