% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{create_keyboard_map}
\alias{create_keyboard_map}
\title{Create Keyboard Map}
\usage{
create_keyboard_map(scale_table, middle_c_octave = 4)
}
\arguments{
\item{scale_table}{an output of \code{create_scale_table}}

\item{middle_c_octave}{octave number for middle C. There are varying
conventions for the octave number for middle C. The default for this
function is 4, but other software can use 3 or even some other number}
}
\value{
the keyboard map. This is a data.table with eight columns:
\itemize{
\item \code{note_number}: the MIDI note number from \code{.NN_MIN} through \code{.NN_MAX}
\item \code{note_name}: the note name
\item \code{ratio_frac}: ratio for the note as a vulgar fraction
\item \code{octave}: the octave number of the note
\item \code{degree}: the scale degree of the note
\item \code{freq}: the frequency in Hz
\item \code{cents}: cents above default MIDI note \code{.NN_MIN}, which has frequency
\code{.FREQ_MIN}.
\item \code{name_12edo}: note name of the key in 12EDO
}
}
\description{
Creates a keyboard map
}
\details{
The function is currently hard-coded to compute the map so that
middle C with frequency \code{.FREQ_MIDDLE_C}is mapped to MIDI note number
\code{.NN_MIDDLE_C} and scale degree 0. With the current constants this is the
same as it is on 12EDO with A440 on note 69. This note is 6000 cents above
MIDI note number 0.
}
\examples{
\dontrun{

hexany_harmonics <- c(1, 3, 5, 7)
hexany_choose = 2
hexany <- create_scale_table(hexany_harmonics, hexany_choose)
print(hexany_keyboard_map <- create_keyboard_map(hexany))
dekany_harmonics <- c(1, 7, 9, 11, 13)
dekany_choose <- 2
dekany <- create_scale_table(dekany_harmonics, dekany_choose)
print(dekany_keyboard_map <- create_keyboard_map(dekany))
print(
  vanilla_keyboard_map <- create_keyboard_map(create_12edo_scale_table())
)
print(
  eikosany_keyboard_map_c3 <-
    create_keyboard_map(create_scale_table(), middle_c_octave = 3)
)
}
}
