% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{drop_last_row}
\alias{drop_last_row}
\title{Remove the octave (last row) from a scale table}
\usage{
drop_last_row(scale_table)
}
\arguments{
\item{scale_table}{A scale table resulting from \code{ps_scale_table}, \code{cps_scale_table} or \code{et_scale_table}}
}
\value{
The scale table without the final periodic note
}
\description{
Remove the octave (last row) from a scale table
}
