% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edo12_names}
\alias{edo12_names}
\title{Note name datasets}
\format{
\subsection{\code{edo12_names}}{

A character vector with 12 elements
}
}
\source{
a music theory class
}
\usage{
edo12_names
}
\description{
For scale tables created using function \code{et_scale_table}, the user must
specify note names. As a convenience, the \code{eikosany} package includes
data sets for \verb{12EDO}, \verb{19EDO}, and \verb{31EDO}.
}
\details{
About the note names: even in 2023, code often breaks when given
characters outside the 7-bit international standard. So we don't even
have flats, let alone half-flats, half-sharps, naturals, or any of the
other symbols alternate tuning theorists have proposed.

So for scales created with equal divisions of the octave, we use a
"/b" for a half-flat, "b" for a flat, "bb" for a double-flat, /#"
for a half-sharp, "#" for a sharp. and "##" for a double-sharp. Thanks
in advance for your understanding.

Note names for 12EDO:
}
\keyword{datasets}
