% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edo19_names}
\alias{edo19_names}
\title{Note names for 19EDO:}
\format{
\subsection{\code{edo19_names}}{

A character vector with 19 elements
}
}
\source{
\url{https://en.wikipedia.org/wiki/19_equal_temperament#Notation}
}
\usage{
edo19_names
}
\description{
Note names for 19EDO:
}
\keyword{datasets}
