% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edo31_names}
\alias{edo31_names}
\title{Note names for 31EDO:}
\format{
\subsection{\code{edo31_names}}{

A character vector with 31 elements
}
}
\source{
\url{https://en.wikipedia.org/wiki/31_equal_temperament#Scale_diagram}
}
\usage{
edo31_names
}
\description{
Note names for 31EDO:
}
\keyword{datasets}
