% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{export_music_object}
\alias{export_music_object}
\title{Export Music Object}
\usage{
export_music_object(music_object, file_name, output_directory = "~/MIDI")
}
\arguments{
\item{music_object}{a \code{gm} music object}

\item{file_name}{character The ".mid" suffix will be supplied}

\item{output_directory}{character, default "~/MIDI". This will be created
if it does not exist.}
}
\value{
the full path to the file
}
\description{
Exports a \code{gm} music object to a MIDI file
}
\examples{
\dontrun{
  eikosany <- cps_scale_table()
  eikosany_map <- keyboard_map(eikosany)
  print(eikosany_music_object <- chord_music_object(
    c(1, 6, 11, 15),
     eikosany_map,
     lowest_note = 40,
     highest_note = 80
  ))
  print(file_path <- export_music_object(
    eikosany_music_object,
    "eikosany_chords"
  ))

  vanilla <- et_scale_table()
  vanilla_map <- keyboard_map(vanilla)
  print(vanilla_music_object <- chord_music_object(
    c(0, 4, 7, 10),
     vanilla_map
  ))
  print(file_path <- export_music_object(
    vanilla_music_object,
    "vanilla_chords"
  ))
}
}
