% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{interval_matrix}
\alias{interval_matrix}
\title{Create Interval Matrix}
\usage{
interval_matrix(scale_table)
}
\arguments{
\item{scale_table}{a scale table from the \code{scale_table} function}
}
\value{
an interval matrix
}
\description{
Creates an interval matrix from a scale table
}
\examples{
\dontrun{

# the defaults yield the 1-3-5-7-9-11 Eikosany
print(eikosany_interval_matrix <-interval_matrix(scale_table()))
}
}
