% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{interval_table}
\alias{interval_table}
\title{Create Interval Table}
\usage{
interval_table(scale_table)
}
\arguments{
\item{scale_table}{a scale table from \code{ps_scale_table},
\code{cps_scale_table}, or \code{et_scale_table}}
}
\value{
an interval table. This is a data.table with seven columns
\itemize{
\item \code{from} name of "from" note
\item \code{from_degree} scale degree of "from" note
\item \code{to} name of "to" note
\item \code{to_degree} scale degree of "to" note
\item \code{ratio} interval as a number
\item \code{ratio_frac} interval as a vulgar fraction (character)
\item \code{ratio_cents} interval in cents
}
}
\description{
Creates an interval table from a scale table
}
\examples{

# default is the 1-3-5-7-9-11 Eikosany
eikosany <- cps_scale_table()
print(eikosany_interval_table <-interval_table(eikosany))
}
