% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{label2harmonics}
\alias{label2harmonics}
\title{Extract unique harmonics from a set of labels}
\usage{
label2harmonics(label, separator)
}
\arguments{
\item{label}{A character vector of labels (e.g., "1x3x5")}

\item{separator}{Character used to separate harmonics (e.g., "x")}
}
\value{
A sorted numeric vector of all unique harmonic factors found in the labels
}
\description{
Extract unique harmonics from a set of labels
}
