% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{label2matrix}
\alias{label2matrix}
\title{Convert a label string to a matrix}
\usage{
label2matrix(label, separator)
}
\arguments{
\item{label}{A character vector of labels (e.g., "1x3x5")}

\item{separator}{Character used to separate harmonics (e.g., "x")}
}
\value{
A matrix where each row is a set of harmonics derived from the label
}
\description{
Convert a label string to a matrix
}
