% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{label2prod}
\alias{label2prod}
\title{Convert a product label to its numerical product}
\usage{
label2prod(label, separator)
}
\arguments{
\item{label}{A character vector of labels (e.g., "1x3x5")}

\item{separator}{Character used to separate harmonics (e.g., "x")}
}
\value{
A numeric vector of products
}
\description{
Convert a product label to its numerical product
}
