% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{matrix2degree}
\alias{matrix2degree}
\title{Map a matrix of notes to their scale degrees}
\usage{
matrix2degree(note_matrix, scale_table)
}
\arguments{
\item{note_matrix}{A matrix where rows are combinations of harmonic factors}

\item{scale_table}{A scale table from \code{ps_scale_table}, \code{cps_scale_table}, or \code{et_scale_table}}
}
\value{
A data.table containing the degree for each note in the matrix
}
\description{
Map a matrix of notes to their scale degrees
}
