% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{matrix2label}
\alias{matrix2label}
\title{Convert a matrix to a label string}
\usage{
matrix2label(matrix, separator)
}
\arguments{
\item{matrix}{A matrix of harmonics}

\item{separator}{Character used to separate harmonics in the label (e.g., "x")}
}
\value{
A character vector of labels
}
\description{
Convert a matrix to a label string
}
