% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{nn2freq}
\alias{nn2freq}
\title{Convert MIDI note number to frequency}
\usage{
nn2freq(note_number)
}
\arguments{
\item{note_number}{MIDI note number (0-127)}
}
\value{
Frequency in Hz
}
\description{
Convert MIDI note number to frequency
}
