% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{period_reduce}
\alias{period_reduce}
\title{Reduce a ratio into a specified period (e.g., an octave)}
\usage{
period_reduce(x, period = 2)
}
\arguments{
\item{x}{A numeric value or vector of ratios}

\item{period}{The period to wrap the ratio into (default is 2 for an octave). Must be greater than 0.}
}
\value{
The reduced ratios, all within the range [1, period)
}
\description{
Reduce a ratio into a specified period (e.g., an octave)
}
