% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{ratio2cents}
\alias{ratio2cents}
\title{Convert a ratio to cents}
\usage{
ratio2cents(ratio)
}
\arguments{
\item{ratio}{Numeric ratio}
}
\value{
Ratio expressed in cents
}
\description{
Convert a ratio to cents
}
