% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{ratio_table}
\alias{ratio_table}
\title{Create a basic ratio table from labels}
\usage{
ratio_table(label, period = 2, root_divisor)
}
\arguments{
\item{label}{A character vector of product labels (e.g., "1x3x5")}

\item{period}{The period to wrap ratios into (default is 2 for an octave)}

\item{root_divisor}{The divisor used to scale the smallest product to 1/1}
}
\value{
A \code{data.table} containing ratio, fractional representation and cents
}
\description{
Create a basic ratio table from labels
}
