% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{render_cps_chords}
\alias{render_cps_chords}
\title{Render the Chords of a CPS}
\usage{
render_cps_chords(scale_table, output_directory)
}
\arguments{
\item{scale_table}{the scale table for the CPS}

\item{output_directory}{chords will be rendered into this folder}
}
\value{
the full path to output_directory
}
\description{
Make WAV files for all chords of a given CPS
}
\examples{
\dontrun{

  hexany_scale_table <- cps_scale_table(
    harmonics = c(1, 3, 5, 7, 9, 11),
    choose = 2,
    root_divisor = 15
  )
  render_cps_chords(hexany_scale_table, "~/Music/hexany_chords")

  eikosany_scale_table <- cps_scale_table(
    harmonics = c(1, 3, 5, 7, 9, 11),
    choose = 3,
    root_divisor = 33
  )
  render_cps_chords(eikosany_scale_table, "~/Music/eikosany_chords")
}

}
