% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{chord_synth}
\alias{chord_synth}
\title{Synthesize a Chord}
\usage{
chord_synth(
  chord,
  signal = "tria",
  duration_sec = 1,
  velocity = 100,
  sample_rate_hz = 48000,
  bit_width = 24
)
}
\arguments{
\item{chord}{a vector of frequencies for the chord}

\item{signal}{The \code{seewave} signal type: "sine", "tria",
"square" or "saw", default = "tria"}

\item{duration_sec}{how many seconds to hold each note,
default = 1}

\item{velocity}{MIDI velocity, default = 100, max is 127}

\item{sample_rate_hz}{sample rate in hz, default = 48000}

\item{bit_width}{bit width of samples, default = 24}
}
\value{
the full path to output_directory
}
\description{
Creates a \code{Wave} object for a given chord
}
\examples{
\dontrun{
justmajor7th <- c(1, 5/4, 3/2, 7/4)
wave <- chord_synth(256*justmajor7th, duration_sec = 10)
tuneR::play(wave)
}

}
