% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{chord_table}
\alias{chord_table}
\title{Create Chord Table}
\usage{
chord_table(scale_table, harmonics = c(1, 3, 5, 7, 9, 11), choose = 4)
}
\arguments{
\item{scale_table}{the scale table to use for note number and name lookup}

\item{harmonics}{a vector of the harmonics to use - defaults to the first
six odd numbers, the harmonics that define the 1-3-5-7-9-11 eikosany.}

\item{choose}{the number of harmonics to choose for each chord -
defaults to 4, to compute tetrads for the eikosany.}
}
\value{
a data.table with five columns:
\itemize{
\item \code{chord_label}: the chord expressed as colon-separated harmonics
\item \code{harmonic_scale_degrees}: the harmonic chord expressed as
colon-separated scale degrees
\item \code{subharmonic_scale_degrees}: the subharmonic chord expressed as
colon-separated scale degrees
}
}
\description{
Creates a chord table
}
\examples{
\dontrun{

# the defaults yield the 1-3-5-7-9-11 Eikosany
eikosany_scale <- scale_table()[["scale_table"]]
print(eikosany_scale)

# the defaults yield the tetrads of the 1-3-5-7-9-11 Eikosany
eikosany_chords <- chord_table(eikosany_scale)
print(eikosany_chords)
}
}
