% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{chord_WAVs}
\alias{chord_WAVs}
\title{Create Chord WAV Files}
\usage{
chord_WAVs(
  chord,
  keyboard_map,
  lowest_note,
  duration_sec,
  velocity = 100,
  output_directory
)
}
\arguments{
\item{chord}{a numeric vector with the scale degrees for the chord}

\item{keyboard_map}{the keyboard map for the scale}

\item{lowest_note}{the lowest MIDI note number to use}

\item{duration_sec}{how long to hold each note}

\item{velocity}{MIDI velocity, default = 100, max is 127}

\item{output_directory}{character, no default; will be created!}
}
\value{
the full path to output_directory
}
\description{
Creates WAV files for inversions of a chord between
two notes
}
