% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{chord_WAVs}
\alias{chord_WAVs}
\title{Create Chord WAV Files}
\usage{
chord_WAVs(
  chord,
  keyboard_map,
  lowest_note = 48,
  highest_note = 84,
  output_directory = "~/Multisample",
  signal = "saw",
  duration_sec = 1,
  velocity = 100,
  sample_rate_hz = 48000,
  bit_width = 24
)
}
\arguments{
\item{chord}{a numeric vector with the scale degrees for the chord}

\item{keyboard_map}{the keyboard map for the scale}

\item{lowest_note}{the lowest MIDI note number to use, default
is 48.}

\item{highest_note}{the highest MIDI note number to use, default
is 84.}

\item{output_directory}{character, default "~/Multisample". This will
be created if it does not exist.}

\item{signal}{the \code{seewave} signal type, default is "saw"}

\item{duration_sec}{how long to hold each note, default = 4}

\item{velocity}{MIDI velocity, default = 100, max is 127}

\item{sample_rate_hz}{sample rate in hz, default = 48000}

\item{bit_width}{bit width of samples, default = 24}
}
\value{
the full path to output_directory
}
\description{
Creates WAV files for inversions of a chord between
two notes
}
\examples{
\dontrun{
eikosany <- cps_scale_table()
eikosany_chords <- cps_chord_table(eikosany)
eikosany_map <- keyboard_map(eikosany)
chord_degrees <- eikosany_chords$degrees
for (i in length(chord_degrees)) {
  chord <- as.numeric(unlist(strsplit(chord_degrees[i], ":")))
  folder_name <-
    paste0("~/eikosany-chords/chord-", gsub(":", "-", chord_degrees[i]))
  print(paste0("generating WAVs in folder ", folder_name))
  chord_WAVs(
    chord,
    keyboard_map = eikosany_map,
    lowest_note = 40,
    highest_note = 80,
    output_directory = folder_name
  )
}
}

}
