% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{create_chord_plot}
\alias{create_chord_plot}
\title{Create Chord Plot}
\usage{
create_chord_plot(chord, title_string, keyboard_map, base_note_number)
}
\arguments{
\item{chord}{a numeric vector with the scale degrees for the chord}

\item{title_string}{A string to use for a plot title. This will usually
be the chord name from a chord table}

\item{keyboard_map}{the keyboard map for the scale}

\item{base_note_number}{must be == 0 modulo 12! The \code{pichor} keyboard
has exactly 24 notes, starts with a "C" and ends with a "B".}
}
\value{
a \code{ggplot2} object that can be printed
}
\description{
Creates a \code{ggplot2} object to plot a chord
}
\examples{
\dontrun{
  eikosany <- create_scale_table()
  eikosany_map <- create_keyboard_map(eikosany)
  print(create_chord_plot(
    c(1, 6, 11, 15),
     "1:3:5:7",
     eikosany_map,
     60
  ))
}
}
