% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{create_chord_table}
\alias{create_chord_table}
\title{Create Chord Table}
\usage{
create_chord_table(scale_table, choose)
}
\arguments{
\item{scale_table}{the scale table to use for note number and name lookup}

\item{choose}{the number of harmonics to choose for each chord}
}
\value{
a data.table with two columns:
\itemize{
\item \code{chord}: the chord expressed as colon-separated harmonics. A
subharmonic chord is prefixed with a "~".
\item \code{degrees}: the chord expressed as colon-separated scale degrees
}
}
\description{
Creates a chord table
}
\examples{
\dontrun{

# the defaults yield the 1-3-5-7-9-11 Eikosany
eikosany_scale <- create_scale_table()
print(eikosany_scale)

# compute the tetrads of the 1-3-5-7-9-11 Eikosany
eikosany_chords <- create_chord_table(eikosany_scale, 4)
print(eikosany_chords)
}
}
