% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edo19_names}
\alias{edo19_names}
\title{Note names for 19EDO}
\format{
\subsection{\code{edo19_names}}{

A character vector with 19 elements
}
}
\source{
\url{https://en.wikipedia.org/wiki/19_equal_temperament#Notation}
}
\usage{
edo19_names
}
\description{
For scale tables created using function \code{et_scale_table}, the user must
specify note names. As a convenience, the \code{eikosany} package includes
data sets for \verb{12EDO}, \verb{19EDO}, and \verb{31EDO}.
}
\details{
Even in 2023, code often breaks when given
characters outside the 7-bit international standard. So we don't even
have flats, let alone half-flats, half-sharps, naturals, or any of the
other symbols alternate tuning theorists have proposed.

So for scales created with equal divisions of the octave, we use a
"/b" for a half-flat, "b" for a flat, "bb" for a double-flat, /#"
for a half-sharp, "#" for a sharp. and "##" for a double-sharp. Thanks
in advance for your understanding.
}
\keyword{datasets}
