% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{et_scale_table}
\alias{et_scale_table}
\title{Create Equal-Tempered Scale Table}
\usage{
et_scale_table(
  note_names = c("C", "C#", "D", "D#", "E", "F", "F#", "G", "G#", "A", "A#", "B"),
  period = 2
)
}
\arguments{
\item{note_names}{a character vector with the names of the notes in the
scale. The default is the names of the standard 12 equal divisions of the
octave.}

\item{period}{The period - default is 2, for an octave}
}
\value{
a \code{data.table} with eight columns:
\itemize{
\item \code{note_name}: the note name (character)
\item \code{ratio}: the ratio that defines the note, as a number between 1 and
\code{period}
\item \code{ratio_frac}: the ratio as a vulgar fraction (character). The ratios
for most EDOs are irrational, so this is an approximation.
\item \code{ratio_cents}: the ratio in cents (hundredths of a semitone)
\item \code{interval_cents}: interval between this note and the previous note
\item \code{degree}: scale degree from zero to (number of notes) - 1
\item \code{key_12EDO}: note name for closest 12EDO note. Note that because
most synthesizers do not allow returning relative to the C in the next
octave up, the offset will be above B for such a note.
\item \code{offset_cents}: offset in cents from \code{key_12EDO}
}

\emph{Note: offsets are meaningless if \code{period} is greater than 2, so in that case they are not computed!}
}
\description{
Creates a scale table for equal divisions of a
specified period.
}
\examples{

print(vanilla <- et_scale_table()) # default is 12EDO, of course

# 19-EDO
# About the note names: even in 2023, code often breaks when given
# characters outside the 7-bit international standard. So we don't even
# have flats, let alone half-flats, half-sharps, naturals, or any of the
# other symbols alternate tuning theorists have proposed.

# So outside of 12EDO, where sharps do double-duty, and product set
# scales, where we can use the product label as a note name, we use a
# "/b" for a half-flat, "b" for a flat, "/#" for a half-sharp, and "#"
# for a sharp. Thanks in advance for your understanding.

nn19 <- c(
  "C",
  "C#",
  "Db",
  "D",
  "D#",
  "Eb",
  "E",
  "E#",
  "F",
  "F#",
  "Gb",
  "G",
  "G#",
  "Ab",
  "A",
  "A#",
  "Bb",
  "B",
  "B#")

print(edo19 <- et_scale_table(nn19))

# 31-EDO
nn31 <- c(
  "C",
  "C/#",
  "C#",
  "Db",
  "D/b",
  "D",
  "D/#",
  "D#",
  "Eb",
  "E/b",
  "E",
  "Fb",
  "E#",
  "F",
  "F/#",
  "F#",
  "Gb",
  "G/b",
  "G",
  "G/#",
  "G#",
  "Ab",
  "A/b",
  "A",
  "A/#",
  "A#",
  "Bb",
  "B/b",
  "B",
  "Cb",
  "B#")
print(edo31 <- et_scale_table(nn31))
}
