% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_functions.R
\name{et_scale_table}
\alias{et_scale_table}
\title{Create Equal-Tempered Scale Table}
\usage{
et_scale_table(
  note_names = c("C", "C#|Db", "D", "D#|Eb", "E", "F", "F#|Gb", "G", "G#|Ab", "A",
    "A#|Bb", "B"),
  period = 2
)
}
\arguments{
\item{note_names}{a character vector with the names of the notes in the
scale. The default is the names of the standard 12 equal divisions of the
octave.}

\item{period}{The period - default is 2, for an octave}
}
\value{
a \code{data.table} with six columns:
\itemize{
\item \code{note_name}: the note name (character)
\item \code{ratio}: the ratio that defines the note, as a number between 1 and
\code{period}
\item \code{ratio_frac}: the ratio as a vulgar fraction (character). The ratios
for most EDOs are irrational, so this is an approximation.
\item \code{ratio_cents}: the ratio in cents (hundredths of a semitone)
\item \code{interval_cents}: interval between this note and the previous note
\item \code{degree}: scale degree from zero to (number of notes) - 1
}
}
\description{
Creates a scale table for equal divisions of a
specified period.
}
\examples{

print(vanilla <- et_scale_table()) # default is 12EDO, of course

# 19-EDO
print(edo19 <- et_scale_table(edo19_names))
# 31-EDO
print(edo31 <- et_scale_table(edo31_names))
}
